refactor(controller): split circuitEmulationPageController into sub-controllers
